Add initial (untested) code
What we have probably won't work, but is sufficient to show that my
design plans should be good.

The HTTP client wrappers expose the underlying error messages, so user
code that inspects those error would have to be rewritten if they change
HTTP clients. Since we merely pass the error along, b2-client can
continue to work backend-agnostically.

The wrapper trait also requires more than the Backblaze API utilizes;
this may be worth moving to a separate crate in the future.

We still need to support the use of any backend rather than only the
ones we've implemented.