Reuse HTTP client between requests.
This makes more sense than creating anew each time, even for cases where
making requests is only an occasional action. The user will have to hold
an authorization token, and we can keep the client in that object; that
allows us to avoid requring user code to manage--or care--about the
client after its initial creation.