Change URLs to localhost for tests
Surf is still sending requests to B2 even with surf-vcr intercepting
them; this is likely upstream's issue 307[1], but I need to debug to be
certain.

This commit works around the problem by using localhost as our server.
The strange thing is that surf does not appear to always send the
requests, and I have not been able to observe any requests when
listening via netcat.

[1]: https://github.com/http-rs/surf/issues/307