Add full support for arbitrary HTTP clients
Error management is a bit messier than I would like, but it keeps the
public API clean.

We need to wrap Surf's error type with something that implements Error
so we can clean up our type bounds. First I want to see why Surf doesn't
impl Error and ensure that it won't cause a problem for us.

As a side-effect, it is now possible to use multiple backends at once;
that's unlikely to be useful -- should we explicitly not support it?