WIP Prepare for bucket functions
Most of the types are finished. There is still quite a bit of validation
code to write and documentation is incomplete.

The bucket module is consistent in using builder() methods; I need to update
account.rs to match.

The validation of lifecycle rules clones all rule paths at least once, and
all conflicts (errors) at least twice. Creating buckets is highly
unlikely to be in anyone's hot path, but we can do better.