Validate capabilities before making API call
Also fixed capability lists provided in tests.

Since we're not always able to validate the capabilities, using the
functions will require checking both Error::B2 and Error::Unauthorized;
we may want to convert all B2 unauthorized responses to an
Error::Unauthorized (at least when the problem is lack of a capability)
to simplify code using the library.