Fully support uploading large files
This commit contains implementations for b2_get_upload_part_url,
b2_upload_part, and b2_finish_large_file.

Instead of a get_upload_part_url function, we've created a
get_upload_authorization function, and use the returned object as an
authorization to upload. Still need to document the difference to the B2
API.