Return non-JSON bodies, headers from HTTP clients
We'll need this to support file downloads.

The hyper implementation could be more efficient in general; since I'll
be replacing it with the low-level clients that hyper uses it's probably
not worth the time.

If we handle headers better (i.e., little to no work) we can simplify
send implementations; we'll see after looking at the way the long-term
default clients handle headers.