Refactor test HTTP client creation
We currently remove secrets identically for all tests, even if a
replacement is only needed for a single test. This allows us to do
one-off replacements, with the added benefit that the change is now a
visible part of the test setup.

Changes needed for all tests (authorization tokens, account IDs, etc.)
continue to be replaced globally.