Support downloading files via a DownloadAuthorization
We cannot use a trait because exposing the client mutably causes
problems with the borrow checker, so I use a DownloadAuth enum.

This does require making the HttpClient cloneable; we could
refcount/etc. but that's an extra indirection on every access in every
function; I'm expecting the use of a DownloadAuth to be rare, and it's
the only time the cost of a clone is paid, so based on my current
assumptions this is better.